+++ Cooperativa/Model/Fabricantes.cs
@@ -10,7 +10,7 @@ namespace Model {
 /*			tmeFabs = new List<TmeFab>();
 			detallesModelosMedidores = new List<DetallesModelosMedidore>();
   */      }
-        public virtual int FabNumero { get; set; }
+        public virtual long FabNumero { get; set; }
         public virtual string FabDescripcion { get; set; }
         public virtual string FabHabilitado { get; set; }
         public virtual long EmpNumero { get; set; }
